Add cli
