+++ backend/app/app/core/config.py
@@ -60,7 +60,7 @@ class Settings(BaseSettings):
     @validator("EMAILS_FROM_NAME")
     def get_project_name(cls, v: Optional[str], values: Dict[str, Any]) -> str:
         if not v:
-            return values["PROJECT_NAME"]
+            return values.get("PROJECT_NAME", "")
         return v
 
     EMAIL_RESET_TOKEN_EXPIRE_HOURS: int = 48
